+++ Infrastructure/Service/HomeApplianceService.cs
@@ -1,6 +1,7 @@
 ﻿using Aplication.Service;
 using Domain.Entitys.Home.SpareParts;
 using Domain.Model;
+using Infrastructure.DataAccess;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -11,11 +12,26 @@ namespace Infrastructure.Service
 {
     public class HomeApplianceService : IHomeAppliance
     {
-        public Task<Response<HomeAppliance>> CreateHomeApplianceAynce(HomeAppliance homeAppliance)
+        private readonly HomeDbContext _homeDbContext;
+
+        public HomeApplianceService(HomeDbContext homeDbContext)
         {
-            throw new NotImplementedException();
+            _homeDbContext = homeDbContext;
         }
 
+        public async Task<Response<HomeAppliance>> CreateHomeApplianceAynce(HomeAppliance homeAppliance)
+        {
+            try
+            {
+                await _homeDbContext.HomeAppliances.AddAsync(homeAppliance);
+                await _homeDbContext.SaveChangesAsync();
+                return new Response<HomeAppliance>(homeAppliance);
+            }
+            catch (Exception ex)
+            {
+                return new Response<HomeAppliance>("Failed to create  homeAppliance: " + ex.Message);
+            }
+        }
         public Task<Response<HomeAppliance>> DeleteHomeApplianceAynce(int Id)
         {
             throw new NotImplementedException();
